move the exception code to its own module because it is used for both client and server
