+++ rodbus/src/lib.rs
@@ -187,6 +187,8 @@ pub const VERSION: &str = env!("CARGO_PKG_VERSION");
 pub mod client;
 /// public constant values related to the Modbus specification
 pub mod constants;
+/// Modbus exception codes
+pub mod exception;
 /// types used to control decoding in the log
 pub mod decode;
 /// error types associated with making requests
