+++ rodbus/src/prelude.rs
@@ -1,7 +1,7 @@
 pub use crate::client::channel::{strategy, Channel, ReconnectStrategy, RequestParam};
 pub use crate::client::{create_handle_and_task, spawn_tcp_client_task};
 pub use crate::decode::*;
-pub use crate::error::details::ExceptionCode;
+pub use crate::exception::ExceptionCode;
 pub use crate::error::*;
 pub use crate::server::handler::{RequestHandler, ServerHandlerMap};
 pub use crate::server::{create_tcp_server_task, spawn_tcp_server_task};
