+++ rodbus/src/server/handler.rs
@@ -1,7 +1,7 @@
 use std::collections::BTreeMap;
 use std::sync::{Arc, Mutex};
 
-use crate::error::details::ExceptionCode;
+use crate::exception::ExceptionCode;
 use crate::types::*;
 
 /// Trait implemented by the user to process requests received from the client
